+++ package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'staringatlights:infinite-scroll',
-  version: '0.9.2',
+  version: '1.0.0',
   // Brief, one-line summary of the package.
   summary: 'Template level infinite scrolling.',
   // URL to the Git repository containing the source code for this package.