+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.0
+- Customizable container and loading template (thanks @VolodymyrTymets)
+- Removed dburles:mongo-collection-instances dependency (thanks @mizsha)
+- Added ecmpascript dependency 
+- Rewritten demo application
+
 ## 0.9.2
 - Removed `dburles:mongo-collection-instances` dependency, thanks @mizsha